hook_ct: remove the non-working with devpts
vzctl mounts devpts with the newinstance option. It doesn't work.

Centos umounts old devpts and tries to mount a new one.
Fedora mounts devtmpfs over /dev.

Instead of this hack, we need to add the newinstance option in
/etc/fstab or /etc/filesystems.

Another reasone to remove it is that this configuration doesn't
supported by CRIU.